rpg 02
Added player and actor components.  The player takes the skin and user directions the actor component animates the player.  Added usewalk and usekey press hooks to get the users keypress and walk the player based on the arrow key directions.  created functionality to map event keypress directions to integers then translate them to positions on the sprite sheet.