rpg 04
finished part one of the rpg.  character can now move around on the screen.  Began work creating a draggable map editor and tile grid for the project.  Added skeleton files.  Created app componen tilePallete and use-draggable hook component.  The app renders the tilepallet and the use draggable maps an image to the screen and allows the user to drag it around the viewport.  The TilePalette component renders the a grid of objects with x and y values then the tile grid is displayed using the x and y values.  Each part of the background spritesheet is mapped to the specific tile using the x,y values from the tile object.