rpg 07
Fixed issue with active tile not being used in map background after switching to a different palatte.  The palatte bg would update and the active tile would show but when clicking on the map it would add the previous background tile.  Added sprite state to map component and updated the active tile background sprite to use the sprite state.   Extracted the rendering into multiple methods for easier understanding and cleaner code.